downgrade to raphael 2.1.1.
This is needed because later versions contain a bug which prevents SVG
gradients to be displayed in case the URL of the page contains a hash (which
is our case, sadly :-/).

This is the commit which introduced the bug:
https://github.com/DmitryBaranovskiy/raphael/commit/424719f7e20e3abfdb787240487b07b7931856a4

there is even a note explaining why it's a problem and how to fix, but
apparently they never fixed it. And I'm not fluent enough in JS to write a
proper bug report :)