Fixing Dockerfile build issues
Due to missing header files the previous Dockerfile could not be built.

The problem occurred when the dependency psutil was being built
resulting in the following error:

No such file or directory #include <linux/types.h>

By adding the apk package linux-headers this problem could be resolved
and the Dockerfile is again buildable.